feat: Add Jest Unit Testing
