+++ config/db/index.ts
@@ -1,10 +1,6 @@
-import express from 'express'
-import router from './routes'
-import cookieParser from 'cookie-parser'
-import cors from 'cors'
+import { app } from '../../app'
 import mongoose from 'mongoose'
 
-const app = express()
 const PORT = parseInt(process.env.PORT!)
 const USERNAME = process.env.DB_USERNAME
 const PASSWORD = process.env.DB_PASSWORD
@@ -12,11 +8,6 @@ const DB = process.env.DB
 const HOSTNAME = process.env.DB_HOSTNAME
 
 const url = `mongodb+srv://${USERNAME}:${PASSWORD}@${HOSTNAME}/${DB}`
-
-app.use(express.json())
-app.use(cookieParser())
-app.use(cors())
-app.use('/', router)
 
 const options = {
     connectTimeoutMS: 30000,
