+++ package.json
@@ -8,7 +8,12 @@
         "tsc": "tsc",
         "start": "nodemon index.ts",
         "lint": "eslint './*/.ts'",
-        "prepare": "husky install"
+        "prepare": "husky install",
+        "test": "jest"
+    },
+    "jest": {
+        "preset": "ts-jest",
+        "testEnvironment": "node"
     },
     "repository": {
         "type": "git",
@@ -22,6 +27,7 @@
     "homepage": "https://github.com/sheinken88/Fast-Delivery-Back#readme",
     "dependencies": {
         "bcrypt": "^5.1.0",
+        "cookie-parser": "^1.4.6",
         "cors": "^2.8.5",
         "dotenv": "^16.3.1",
         "express": "^4.18.2",
@@ -36,8 +42,10 @@
         "@types/cookie-parser": "^1.4.3",
         "@types/cors": "^2.8.13",
         "@types/express": "^4.17.17",
+        "@types/jest": "^29.5.3",
         "@types/jsonwebtoken": "^9.0.2",
         "@types/node": "^20.4.2",
+        "@types/supertest": "^2.0.12",
         "@typescript-eslint/eslint-plugin": "^5.52.0",
         "@typescript-eslint/parser": "^5.52.0",
         "eslint": "^8.47.0",
@@ -45,7 +53,10 @@
         "eslint-config-standard-with-typescript": "^37.0.0",
         "eslint-plugin-prettier": "^5.0.0",
         "husky": "^8.0.3",
+        "jest": "^29.6.2",
         "lint-staged": "^14.0.0",
-        "prettier": "^3.0.1"
+        "prettier": "^3.0.1",
+        "supertest": "^6.3.3",
+        "ts-jest": "^29.1.1"
     }
 }
